fixed bugs, fixed UI skill, added paddle ball, will change enemy attack

+++ Assets/Scripts/GamePlay/Player/PlayerCount.cs
@@ -1,5 +1,4 @@
 using DungTran31.GamePlay.Enemy;
-using System;
 using TMPro;
 using UnityEngine;
 
@@ -10,8 +9,8 @@ namespace DungTran31.GamePlay.Player
         [SerializeField] private TextMeshProUGUI killRemainingCountText;
         [SerializeField] private GameObject bossPrefab;
         [SerializeField] private int targetKillCount;
+        [SerializeField] private PlayerBall playerBall;
 
-        public static event Action OnTargetKillCountReached;
         public int KillCount { get; private set; }
         private bool bossSpawned = false;
 
@@ -28,8 +27,11 @@ namespace DungTran31.GamePlay.Player
 
         private void IncrementKillCount(EnemyHealth.EnemyDeathEventArgs args)
         {
-            KillCount++; // Increment the kill count
-            UpdateKillCountUI(); // Update the UI
+            if (KillCount < targetKillCount)
+            {
+                KillCount++; // Increment the kill count
+                UpdateKillCountUI(); // Update the UI
+            }
         }
 
         // Updates the kill count UI text
@@ -38,23 +40,28 @@ namespace DungTran31.GamePlay.Player
             if (killRemainingCountText != null)
             {
                 int remainingKillCount = targetKillCount - KillCount;
-                print("Remaining kill count: " + remainingKillCount);
-                if(remainingKillCount <= 0)
+
+                if (remainingKillCount <= 0)
                 {
-                    remainingKillCount = 0;
                     if (!bossSpawned)
                     {
                         // Spawn the boss when the player reaches the target kill count
                         Instantiate(bossPrefab, Vector3.zero, Quaternion.identity);
-                        OnTargetKillCountReached?.Invoke();
+                        bossSpawned = true;
                     }
-                    bossSpawned = true;
+                    ActivatePlayerBall(); // Activate the player ball
                 }
                 killRemainingCountText.text = remainingKillCount.ToString();
             }
         }
 
-        
+        private void ActivatePlayerBall()
+        {
+            if (playerBall != null)
+            {
+                playerBall.gameObject.SetActive(true);
+                playerBall.ActivatePlayerBall();
+            }
+        }
     }
-
 }
